implementation/requirements.txt
streamlit>=1.35.0
langchain>=0.2.0
langchain-community>=0.2.0
langchain-openai>=0.1.0
langchain-groq>=0.1.0
langchain-huggingface>=0.0.3
langchain-text-splitters>=0.2.0
faiss-cpu>=1.8.0
pypdf>=4.0.0
sentence-transformers>=3
python-dotenv>=1.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
email-validator>=2.0.0
jinja2>=3.1.0
